% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuschat.R
\name{census_chat_shiny_basic}
\alias{census_chat_shiny_basic}
\title{Basic Shiny Chat Interface}
\usage{
census_chat_shiny_basic(
  provider = "anthropic/claude-sonnet-4-5-20250929",
  port = NULL,
  launch.browser = TRUE
)
}
\description{
Basic Shiny Chat Interface
}
\keyword{internal}
