% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get Census API Key}
\usage{
get_api_key()
}
\value{
The API key as a string. Throws an error if no key is found.
}
\description{
Get Census API Key
}
\keyword{internal}
