% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuschat.R
\name{resolve_state_fips}
\alias{resolve_state_fips}
\title{Resolve state name/abbreviation to FIPS code}
\usage{
resolve_state_fips(state)
}
\description{
Resolve state name/abbreviation to FIPS code
}
\keyword{internal}
