% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_fetch_data}
\alias{tool_fetch_data}
\title{Tool: Fetch Census Data using tidycensus}
\usage{
tool_fetch_data()
}
\value{
An ellmer \code{\link[=tool]{tool()}} definition.
}
\description{
Tool: Fetch Census Data using tidycensus
}
\keyword{internal}
