% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_get_place_tracts}
\alias{tool_get_place_tracts}
\title{Tool: Get Tract Data for a Place}
\usage{
tool_get_place_tracts()
}
\value{
An ellmer \code{\link[=tool]{tool()}} definition.
}
\description{
Tool: Get Tract Data for a Place
}
\keyword{internal}
