% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_get_variables}
\alias{tool_get_variables}
\title{Tool: Get Common Variables}
\usage{
tool_get_variables()
}
\value{
An ellmer \code{\link[=tool]{tool()}} definition.
}
\description{
Tool: Get Common Variables
}
\keyword{internal}
