% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tool_resolve_fips}
\alias{tool_resolve_fips}
\title{Tool: Resolve FIPS Codes}
\usage{
tool_resolve_fips()
}
\value{
An ellmer \code{\link[=tool]{tool()}} definition.
}
\description{
Tool: Resolve FIPS Codes
}
\keyword{internal}
